backend/freelanco/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
idna==2.9
oauthlib==3.1.0
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
urllib3==1.25.8
